Correction de légers détails.
Cela n'affecte en rien le déroulement des tests, mais pour certaines opérations, ça affichait des int au lieu de float...

+++ src/main/java/fr/ensimag/deca/tree/ReadFloat.java
@@ -19,7 +19,7 @@ public class ReadFloat extends AbstractReadExpr {
     @Override
     public Type verifyExpr(DecacCompiler compiler, EnvironmentExp localEnv,
             ClassDefinition currentClass) throws ContextualError {
-                Type type = new FloatType(compiler.getSymbolTable().create("int"));
+                Type type = new FloatType(compiler.getSymbolTable().create("float"));
                 this.setType(type);
                 return type;
     }
